+++ src/parlant/core/services/tools/service_registry.py
@@ -31,6 +31,7 @@ from parlant.core.nlp.service import NLPService
 from parlant.core.persistence.document_database_helper import DocumentStoreMigrationHelper
 from parlant.core.services.tools.openapi import OpenAPIClient
 from parlant.core.services.tools.plugins import PluginClient
+from parlant.core.services.tools.mcp_service import MCPToolClient
 from parlant.core.tools import LocalToolService, ToolService
 from parlant.core.common import ItemNotFoundError, Version, UniqueId
 from parlant.core.persistence.common import ObjectId
@@ -41,7 +42,7 @@ from parlant.core.persistence.document_database import (
 )
 
 
-ToolServiceKind = Literal["openapi", "sdk", "local"]
+ToolServiceKind = Literal["openapi", "sdk", "local", "mcp"]
 
 
 class ServiceRegistry(ABC):
@@ -137,11 +138,15 @@ class ServiceDocumentRegistry(ServiceRegistry):
     def _cast_to_specific_tool_service_class(
         self,
         service: ToolService,
-    ) -> OpenAPIClient | PluginClient:
-        if isinstance(service, OpenAPIClient):
-            return service
-        else:
-            return cast(PluginClient, service)
+    ) -> OpenAPIClient | PluginClient | MCPToolClient:
+        if not (
+            isinstance(service, OpenAPIClient)
+            or isinstance(service, PluginClient)
+            or isinstance(service, MCPToolClient)
+        ):
+            raise ValueError("Unsupported ToolService class.")
+
+        return service
 
     async def _document_loader(self, doc: BaseDocument) -> Optional[_ToolServiceDocument]:
         if doc["version"] == "0.1.0":
@@ -210,14 +215,26 @@ class ServiceDocumentRegistry(ServiceRegistry):
         name: str,
         service: ToolService,
     ) -> _ToolServiceDocument:
+        kind: ToolServiceKind
+
+        if isinstance(service, OpenAPIClient):
+            kind = "openapi"
+            url = service.server_url
+        elif isinstance(service, PluginClient):
+            kind = "sdk"
+            url = service.url
+        elif isinstance(service, MCPToolClient):
+            kind = "mcp"
+            url = service.url
+        else:
+            raise ValueError("Unsupported ToolService class.")
+
         return _ToolServiceDocument(
             id=ObjectId(name),
             version=self.VERSION.to_string(),
             name=name,
-            kind="openapi" if isinstance(service, OpenAPIClient) else "sdk",
-            url=service.server_url
-            if isinstance(service, OpenAPIClient)
-            else cast(PluginClient, service).url,
+            kind=kind,
+            url=url,
             source=self._service_sources.get(name) if isinstance(service, OpenAPIClient) else None,
         )
 
@@ -231,6 +248,13 @@ class ServiceDocumentRegistry(ServiceRegistry):
             )
         elif document["kind"] == "sdk":
             return PluginClient(
+                url=document["url"],
+                event_emitter_factory=self._event_emitter_factory,
+                logger=self._logger,
+                correlator=self._correlator,
+            )
+        elif document["kind"] == "mcp":
+            return MCPToolClient(
                 url=document["url"],
                 event_emitter_factory=self._event_emitter_factory,
                 logger=self._logger,
@@ -259,13 +283,22 @@ class ServiceDocumentRegistry(ServiceRegistry):
                 openapi_json = await self._get_openapi_json_from_source(source)
                 service = OpenAPIClient(server_url=url, openapi_json=openapi_json)
                 self._service_sources[name] = source
-            else:
+            elif kind == "mcp":
+                service = MCPToolClient(
+                    url=url,
+                    event_emitter_factory=self._event_emitter_factory,
+                    logger=self._logger,
+                    correlator=self._correlator,
+                )
+            elif kind == "sdk":
                 service = PluginClient(
                     url=url,
                     event_emitter_factory=self._event_emitter_factory,
                     logger=self._logger,
                     correlator=self._correlator,
                 )
+            else:
+                raise ValueError(f"Unsupported ToolService kind: {kind}")
 
             if name in self._running_services:
                 await (