+++ CHANGELOG.md
@@ -7,7 +7,8 @@ All notable changes to Parlant will be documented here.
 - Use fast fluid draft if no utterance found
 - Add logger to LoadedContext (accessible via engine hooks)
 - Add ability to update session mode directly from API
-- Added and improved support for tool parameter types: date, datetime, optional (bug fix) and lists of other allowed types
+- Add an improved support for tool parameter types: date, datetime, optional (bug fix) and lists of other allowed types
+- Add integration with MCP servers for tools (currently only tools, limited MCP types support)
 
 ## [2.2.0] - 2025-05-20